[test]: ModalBtn.jsx 스토리북 파일 추가 및 다른 버튼 컴포넌트 삭제
