+++ src/common/components/button/index.jsx
@@ -1,4 +1,1 @@
-export { default as ExtraLargeBtn } from "@common-components/button/ExtraLargeBtn"
-export { default as LargeBtn } from "@common-components/button/LargeBtn"
-export { default as SmallBtn } from "@common-components/button/SmallBtn"
-export { default as ExtraSmallBtn } from "@common-components/button/ExtraSmallBtn"
+export { default as ModalBtn } from "@common-components/button/ModalBtn"
